ViewOptions and vertical or horizontal layout
