+++ src/components/TranscriptionViewer-v3/Index.jsx
@@ -1,10 +1,9 @@
 import React, { PropTypes } from 'react';
 import { connect } from 'react-redux';
-import apiClient from 'panoptes-client/lib/api-client.js';
 import { fetchSubject, setView } from '../../actions/transcription-viewer-v3.js';
-import { Utility, KEY_CODES } from '../../tools/Utility.js';
 import * as status from '../../constants/status.js';
 
+import ControlPanel from './ControlPanel.jsx';
 import SVGViewer from './SVGViewer.jsx';
 import SVGImage from './SVGImage.jsx';
 import SVGAggregatedText from './SVGAggregatedText.jsx';
@@ -12,37 +11,14 @@ import SVGAggregatedText from './SVGAggregatedText.jsx';
 const DEFAULT_SVGVIEWER_WIDTH = 480;
 const DEFAULT_SVGVIEWER_HEIGHT = 640;
 
-const SHOWAGGREGATIONS_NONE = 'show-aggregations-none';
-const SHOWAGGREGATIONS_OVERLAY = 'show-aggregations-overlay';
-const SHOWAGGREGATIONS_FULL = 'show-aggregations-full';
-const SHOWAGGREGATIONS_CENSORED = 'show-aggregations-censored';
-
 class Index extends React.Component {
   constructor(props) {
     super(props);
-    this.inputSubjectID = null;
-    this.execFetchSubject = this.execFetchSubject.bind(this);
-    
-    this.inputScale = null;
-    this.inputTranslateX = null;
-    this.inputTranslateY = null;
-    this.inputRotate = null;
-    this.updateTransform = this.updateTransform.bind(this);
     this.imageHasLoaded = this.imageHasLoaded.bind(this);
     
     this.state = {
-      scale: 1,
-      translateX: 0,
-      translateY: 0,
-      rotate: 0,
       loadedImage: { width: 0, height: 0 },
-      showAggregations: SHOWAGGREGATIONS_NONE,
     };
-  }
-  
-  execFetchSubject() {
-    const subjectId = this.inputSubjectID.value;
-    this.props.dispatch(fetchSubject(subjectId));
   }
   
   imageHasLoaded(img) {
@@ -63,186 +39,71 @@ class Index extends React.Component {
     return (
       <div className="transcription-viewer-v3">
         
-        <div className="control-panel">
-          <div className="subjectID-subpanel">
-            <input type="text" ref={(ele) => { this.inputSubjectID = ele; }}
-              placeholder="Panoptes Subject ID, e.g. 1274999"
-              onKeyPress={(e) => {
-                if (Utility.getKeyCode(e) === KEY_CODES.ENTER) {
-                  this.execFetchSubject();
+        <ControlPanel />
+        
+        <div className={'data-panel' + ((this.props.viewOptions && this.props.viewOptions.layout === 'vertical') ? ' vertical' : '') }>
+          {(this.props.subjectData && this.props.subjectData.locations && this.props.subjectData.locations.length > 0)
+            ? <div className="data-subpanel">
+                <SVGViewer
+                  className="show-aggregations-none"
+                  width={DEFAULT_SVGVIEWER_WIDTH}
+                  height={DEFAULT_SVGVIEWER_HEIGHT}
+                >
+                {this.props.subjectData.locations.map((loc, locIndex) => {
+                  return <SVGImage key={'image-'+locIndex} src={loc["image/jpeg"]} onLoad={this.imageHasLoaded} />;
+                })}
+
+                {(this.props.aggregationsData)
+                  ? this.props.aggregationsData.map((agg, index) => {
+                    return (
+                      <SVGAggregatedText
+                        key={'aggtext_' + agg.startX + '_' + agg.startY}
+                        offsetX={this.state.loadedImage.width * -0.5}
+                        offsetY={this.state.loadedImage.height * -0.5}
+                        aggregation={agg}
+                        index={index}
+                      />
+                    );
+                  })
+                  : null
                 }
-              }}
-            />
-            <button onClick={this.execFetchSubject}>&raquo;</button>
-          </div>
-          <div className="status-subpanel">
-            {(this.props.subjectID)
-              ? (<p>
-                  <button onClick={this.goToPrevPage.bind(this)}>&laquo;</button>
-                  <span>Subject ID {this.props.subjectID}</span>
-                  <button onClick={this.goToNextPage.bind(this)}>&raquo;</button>
-                </p>)
-              : null
-            }
-            
-            {(() => {
-              switch (this.props.subjectStatus) {
-                case status.STATUS_IDLE:
-                  return <p>Type in a Panoptes Subject ID to search!</p>;
-                case status.STATUS_LOADING:
-                  return <p>Looking for Subject...</p>;
-                case status.STATUS_READY:
-                  return <p>Subject ready.</p>;
-                case status.STATUS_ERROR:
-                  return <p>WHOOPS - Something went wrong!</p>;
-              }
-              return null;
-            })()}
-            
-            {(() => {
-              if (this.props.subjectStatus !== status.STATUS_READY) return null;
-              
-              switch (this.props.subjectStatus) {
-                case status.STATUS_LOADING:
-                  return <p>Looking for Aggregations...</p>;
-                case status.STATUS_READY:
-                  return (
-                    <p>
-                      Aggregations ready.
-                      <select onChange={this.setShowAggregations.bind(this)} value={this.state.showAggregations}>
-                        <option value={SHOWAGGREGATIONS_NONE}>None</option>
-                        <option value={SHOWAGGREGATIONS_OVERLAY}>Overlay</option>
-                        <option value={SHOWAGGREGATIONS_FULL}>Full</option>
-                        <option value={SHOWAGGREGATIONS_CENSORED}>Censored by the FBI</option>
-                      </select>
-                    </p>
-                  );
-                case status.STATUS_ERROR:
-                  return <p>No Aggregations, sorry.</p>;
-              }
-              return null;
-            })()}
-          </div>
-          
-          <div className="viewControl-subpanel">
-            <div className="row">
-              <label>Scale</label>
-              <span className="data">
-                <input
-                  value={this.props.scale}
-                  ref={(itm) => { this.inputScale = itm; }}
-                  onChange={this.updateTransform}
-                  type="number"
-                  step="0.1"
-                  min="0.1" />
-              </span>
-            </div>
-            <div className="row">
-              <label>Translate (x,y)</label>
-              <span className="data">
-                <input
-                  value={this.props.translateX}
-                  ref={(itm) => { this.inputTranslateX = itm; }}
-                  onChange={this.updateTransform}
-                  type="number"
-                  step="10" />
-                <input
-                  value={this.props.translateY}
-                  ref={(itm) => { this.inputTranslateY = itm; }}
-                  onChange={this.updateTransform}
-                  type="number"
-                  step="10" />
-              </span>
-            </div>
-            <div className="row">
-              <label>Rotate (deg)</label>
-              <span className="data">
-                <input
-                  value={this.props.rotate}
-                  ref={(itm) => { this.inputRotate = itm; }}
-                  onChange={this.updateTransform} 
-                  type="number"
-                  step="15" />
-              </span>
-            </div>
-          </div>
-          
-          {(this.props.subjectData && this.props.subjectData.metadata)
-            ? (() => {
-                let metadata = [];
-                for (let m in this.props.subjectData.metadata) {
-                  metadata.push(<div className="row" key={m}><label>{m}</label><span className="data">{this.props.subjectData.metadata[m]}</span></div>);
+
+                </SVGViewer>
+              </div>
+            : null
+          }
+
+          {(this.props.subjectData && this.props.subjectData.locations && this.props.subjectData.locations.length > 0)
+            ? <div className="data-subpanel">
+                {console.log('x'.repeat(40), this.props.viewOptions)}
+                <SVGViewer
+                  className="show-aggregations-full"
+                  width={DEFAULT_SVGVIEWER_WIDTH}
+                  height={DEFAULT_SVGVIEWER_HEIGHT}
+                >
+                {this.props.subjectData.locations.map((loc, locIndex) => {
+                  return <SVGImage key={'image-'+locIndex} src={loc["image/jpeg"]} />;
+                })}
+
+                {(this.props.aggregationsData)
+                  ? this.props.aggregationsData.map((agg, index) => {
+                    return (
+                      <SVGAggregatedText
+                        key={'aggtext_' + agg.startX + '_' + agg.startY}
+                        offsetX={this.state.loadedImage.width * -0.5}
+                        offsetY={this.state.loadedImage.height * -0.5}
+                        aggregation={agg}
+                        index={index}
+                      />
+                    );
+                  })
+                  : null
                 }
-                return (
-                  <div className="metadata-subpanel">
-                    {metadata}
-                  </div>
-                );
-              })()
+                </SVGViewer>
+              </div>
             : null
           }
         </div>
-        
-        {(this.props.subjectData && this.props.subjectData.locations && this.props.subjectData.locations.length > 0)
-          ? <div className="data-panel">
-              <SVGViewer
-                className={this.state.showAggregations}
-                width={DEFAULT_SVGVIEWER_WIDTH}
-                height={DEFAULT_SVGVIEWER_HEIGHT}
-              >
-              {this.props.subjectData.locations.map((loc, locIndex) => {
-                return <SVGImage key={'image-'+locIndex} src={loc["image/jpeg"]} onLoad={this.imageHasLoaded} />;
-              })}
-
-              {(this.props.aggregationsData)
-                ? this.props.aggregationsData.map((agg, index) => {
-                  return (
-                    <SVGAggregatedText
-                      key={'aggtext_' + agg.startX + '_' + agg.startY}
-                      offsetX={this.state.loadedImage.width * -0.5}
-                      offsetY={this.state.loadedImage.height * -0.5}
-                      aggregation={agg}
-                      index={index}
-                    />
-                  );
-                })
-                : null
-              }
-
-              </SVGViewer>
-            </div>
-          : null
-        }
-        
-        {(this.props.subjectData && this.props.subjectData.locations && this.props.subjectData.locations.length > 0)
-          ? <div className="data-panel">
-              <SVGViewer
-                className="show-aggregations-full"
-                width={DEFAULT_SVGVIEWER_WIDTH}
-                height={DEFAULT_SVGVIEWER_HEIGHT}
-              >
-              {this.props.subjectData.locations.map((loc, locIndex) => {
-                return <SVGImage key={'image-'+locIndex} src={loc["image/jpeg"]} />;
-              })}
-
-              {(this.props.aggregationsData)
-                ? this.props.aggregationsData.map((agg, index) => {
-                  return (
-                    <SVGAggregatedText
-                      key={'aggtext_' + agg.startX + '_' + agg.startY}
-                      offsetX={this.state.loadedImage.width * -0.5}
-                      offsetY={this.state.loadedImage.height * -0.5}
-                      aggregation={agg}
-                      index={index}
-                    />
-                  );
-                })
-                : null
-              }
-              </SVGViewer>
-            </div>
-          : null
-        }
         
         {(this.props.aggregationsData)
           ? <div className="examination-panel">
@@ -264,7 +125,7 @@ class Index extends React.Component {
           : null
         }
         
-        {(this.props.currentAggregation !== null && this.props.aggregationsData && this.props.aggregationsData[this.props.currentAggregation])
+        {/*(this.props.currentAggregation !== null && this.props.aggregationsData && this.props.aggregationsData[this.props.currentAggregation])
           ? <div className="aggregation-panel">
               <div className="aggregated">{this.props.aggregationsData[this.props.currentAggregation].text}</div>
               {(() => {
@@ -277,7 +138,7 @@ class Index extends React.Component {
               })()}
             </div>
           : null
-        }
+        */}
         
       </div>
     );
@@ -289,80 +150,29 @@ class Index extends React.Component {
     //this.props.dispatch(fetchSubject('671183'));  //AnnoTate
     this.props.dispatch(fetchSubject('671189'));  //AnnoTate with some... funky fresh text directions
   }
-  
-  updateTransform(e) {
-    this.props.dispatch(setView(
-      parseFloat(this.inputRotate.value),
-      parseFloat(this.inputScale.value),
-      parseFloat(this.inputTranslateX.value),
-      parseFloat(this.inputTranslateY.value),
-    ));
-  }
-  
-  goToNextPage() {
-    if (this.props.subjectID === null) return;  //Can't use (!this.props.subjectID) since '0' is a valid subjectID.
-    try {
-      const targetSubjectID = parseInt(this.props.subjectID) + 1;
-      this.props.dispatch(fetchSubject(targetSubjectID.toString()));
-    } catch (err) {}
-  }
-  
-  goToPrevPage() {
-    if (this.props.subjectID === null) return;  //Can't use (!this.props.subjectID) since '0' is a valid subjectID.
-    try {
-      const targetSubjectID = parseInt(this.props.subjectID) - 1;
-      this.props.dispatch(fetchSubject(targetSubjectID.toString()));
-    } catch (err) {}
-  }
-  
-  setShowAggregations(e) {
-    const val = e.target.options[e.target.selectedIndex].value;
-
-    this.setState({
-      showAggregations: val
-    });
-  }
 }
 
 Index.propTypes = {
-  subjectID: PropTypes.string,
   subjectData: PropTypes.object,
-  subjectStatus: PropTypes.string,
   aggregationsData: PropTypes.array,
-  aggregationsStatus: PropTypes.string,
-  currentAggregation: PropTypes.number,
-  rotate: PropTypes.number,
-  scale: PropTypes.number,
-  translateX: PropTypes.number,
-  translateY: PropTypes.number,
+  viewOptions: PropTypes.object,
 };
 
 Index.defaultProps = {
-  subjectID: null,
   subjectData: null,
-  subjectStatus: null,
   aggregationsData: null,
-  aggregationsStatus: null,
-  currentAggregation: null,
-  rotate: 0,
-  scale: 1,
-  translateX: 0,
-  translateY: 0,
+  viewOptions: {
+    layout: 'horizontal'
+  },
 };
 
 const mapStateToProps = (state) => {
   const store = state.transcriptionViewerV3;
+  console.log('!'.repeat(40), store.viewOptions)
   return {
-    subjectID: store.subjectID,
     subjectData: store.subjectData,
-    subjectStatus: store.subjectStatus,
     aggregationsData: store.aggregationsData,
-    aggregationsStatus: store.aggregationsStatus,
-    currentAggregation: store.currentAggregation,
-    rotate: store.viewRotate,
-    scale: store.viewScale,
-    translateX: store.viewTranslateX,
-    translateY: store.viewTranslateY,
+    viewOptions: store.viewOptions,
   };
 };
