+++ src/index.tpl.html
@@ -6,6 +6,7 @@
     <title>Zooniverse React Transcriber</title>
     <meta name="description" content="A viewer for Zooniverse Transcription projects">
     <meta name="viewport" content="width=device-width, initial-scale=1">
+    <script src="https://use.fontawesome.com/d08f36d12e.js"></script>
   </head>
   <body>
     <div id="root"></div>
