+++ src/styles/components/app.styl
@@ -45,3 +45,17 @@
     &.active,
     &:hover
       background-color: $tertiary
+
+/*  Common Elements
+ */
+
+.button
+  display: inline-block
+  margin: 0.2em 0.5em
+  padding: 0.2em 1em
+  border: 1px solid $primary
+  color: $primary
+  background: $secondary
+  vertical-align: middle
+  border-radius: 1em
+  cursor: pointer
