+++ src/styles/global/root.styl
@@ -1,6 +1,7 @@
 $primary = #939
 $secondary = #fff
 $tertiary = #336
+$quaternary = #eec
 
 :root
   font-family Arial, Helvetica, sans-serif
